Excluded attributive binders
