+++ src/STrain.CQS.Http/RequestSending/Binders/Attributive/AttributiveMethodProvider.cs
