+++ src/STrain.CQS.Http/RequestSending/Binders/IMethodBinder.cs
