--- src/STrain.CQS.Http/RequestSending/Providers/IMethodProvider.cs
@@ -1,8 +1,0 @@
-﻿namespace STrain.CQS.Http.RequestSending.Providers
-{
-    public interface IMethodProvider
-    {
-        HttpMethod GetMethod<TRequest>()
-            where TRequest : IRequest;
-    }
-}
