--- src/STrain.CQS.Http/RequestSending/Providers/IParameterProvider.cs
@@ -1,8 +1,0 @@
-﻿namespace STrain.CQS.Http.RequestSending.Providers
-{
-    public interface IParameterProvider
-    {
-        Task SetParametersAsync<TRequest>(HttpRequestMessage message, TRequest request, CancellationToken cancellationToken)
-            where TRequest : IRequest;
-    }
-}
