--- src/STrain.CQS.Http/RequestSending/Providers/IPathProvider.cs
@@ -1,8 +1,0 @@
-﻿namespace STrain.CQS.Http.RequestSending.Providers
-{
-    public interface IPathBinder
-    {
-        string GetPath<TRequest>(TRequest request)
-            where TRequest : IRequest;
-    }
-}
