+++ src/STrain.CQS.Http/STrain.CQS.Http.csproj
@@ -7,6 +7,12 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Remove="RequestSending\Binders\Attributive\**" />
+    <EmbeddedResource Remove="RequestSending\Binders\Attributive\**" />
+    <None Remove="RequestSending\Binders\Attributive\**" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.Http.Abstractions" Version="2.2.0" />
     <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="6.0.2" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
