+++ src/STrain.CQS.NetCore.LigtInject/HttpRequestSenderBuilderExtensions.cs
@@ -2,8 +2,6 @@
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using STrain.CQS.Http.RequestSending;
-using STrain.CQS.Http.RequestSending.Providers;
-using STrain.CQS.Http.RequestSending.Providers.Attributive;
 using STrain.CQS.NetCore.Builders;
 using STrain.CQS.NetCore.ErrorHandling;
 using STrain.CQS.Senders;
@@ -12,61 +10,6 @@ namespace STrain.CQS.NetCore.LigtInject
 {
     public static class HttpRequestSenderBuilderExtensions
     {
-        public static HttpRequestSenderBuilder UseGenericDefaults(this HttpRequestSenderBuilder builder)
-        {
-            builder.UseDefaults();
-            builder.UseGenericErrorHandler();
-
-            return builder;
-        }
-
-        public static HttpRequestSenderBuilder UseDefaults(this HttpRequestSenderBuilder builder)
-        {
-            builder.UseAttributivePathProvider();
-            builder.UseAttributeMethodProvider();
-            builder.UseAttributiveParameterProviders();
-            builder.UseResponseReaders();
-
-            builder.UseDefaultErrorHandler();
-
-            return builder;
-        }
-
-        public static HttpRequestSenderBuilder UseAttributivePathProvider(this HttpRequestSenderBuilder builder)
-        {
-            builder.Builder.Host.ConfigureContainer<IServiceRegistry>((_, registry) => registry.RegisterTransient<IPathBinder>(factory => new AttributivePathProvider(factory.GetInstance<IOptionsSnapshot<HttpRequestSenderOptions>>().Get(builder.Key).Path, factory.GetInstance<ILogger<AttributivePathProvider>>()), builder.Key));
-            return builder;
-        }
-        public static HttpRequestSenderBuilder UsePathProvider<TPathProvider>(this HttpRequestSenderBuilder builder)
-            where TPathProvider : class, IPathBinder
-        {
-            builder.Builder.Host.ConfigureContainer<IServiceRegistry>((_, registry) => registry.RegisterTransient<IPathBinder, TPathProvider>(builder.Key));
-            return builder;
-        }
-
-        public static HttpRequestSenderBuilder UseAttributeMethodProvider(this HttpRequestSenderBuilder builder) => builder.UseMethodProvider<AttributiveMethodProvider>();
-        public static HttpRequestSenderBuilder UseMethodProvider<TMethodProvider>(this HttpRequestSenderBuilder builder)
-            where TMethodProvider : class, IMethodProvider
-        {
-            builder.Builder.Host.ConfigureContainer<IServiceRegistry>((_, registry) => registry.RegisterTransient<IMethodProvider, TMethodProvider>(builder.Key));
-            return builder;
-        }
-
-        public static HttpRequestSenderBuilder UseAttributiveParameterProviders(this HttpRequestSenderBuilder builder) => builder.UseParameterProviders<AttributiveHeaderParameterProvider, AttributiveQueryParameterProvider, AttributiveBodyParameterProvider>();
-        public static HttpRequestSenderBuilder UseParameterProviders<THeaderParameterProvider, TQueryParameterProvider, TBodyParameterProvider>(this HttpRequestSenderBuilder builder)
-            where THeaderParameterProvider : class, IParameterProvider
-            where TQueryParameterProvider : class, IParameterProvider
-            where TBodyParameterProvider : class, IParameterProvider
-        {
-            builder.Builder.Host.ConfigureContainer<IServiceRegistry>((_, registry) =>
-            {
-                registry.RegisterTransient<IParameterProvider, THeaderParameterProvider>($"{builder.Key}.header");
-                registry.RegisterTransient<IParameterProvider, TQueryParameterProvider>($"{builder.Key}.query");
-                registry.RegisterTransient<IParameterProvider, TBodyParameterProvider>($"{builder.Key}.body");
-            });
-            return builder;
-        }
-
         public static HttpRequestSenderBuilder UseResponseReaders(this HttpRequestSenderBuilder builder) => builder.UseResponseReaders(register => register.UseDefaults());
         public static HttpRequestSenderBuilder UseResponseReaders(this HttpRequestSenderBuilder builder, Action<ResponseReadersRegister> registrate)
         {
