+++ src/STrain.CQS.Test.Function/Drivers/ApiDriver.cs
@@ -5,6 +5,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Moq;
 using STrain.CQS.Http.RequestSending;
+using STrain.CQS.Http.RequestSending.Binders;
 using STrain.CQS.Http.RequestSending.Providers;
 using STrain.CQS.Senders;
 using STrain.Extensions.Testing.Drivers;
@@ -51,6 +52,7 @@ namespace STrain.CQS.Test.Function.Drivers
                                 scope.GetInstance<IRouteBinder>(key),
                                 scope.GetInstance<IMethodBinder>(key),
                                 scope.GetInstance<IQueryParameterBinder>(key),
+                                scope.GetInstance<IHeaderParameterBinder>(key),
                                 scope.GetInstance<IResponseReaderProvider>(key),
                                 scope.GetInstance<IRequestErrorHandler>(key),
                                 scope.GetInstance<ILogger<HttpRequestSender>>());
