+++ src/STrain.CQS.Test.Unit/Http/RequestSending/AttributiveBodyParameterProviderTest.cs
@@ -1,7 +1,6 @@
 ﻿using AutoFixture;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
-using STrain.CQS.Http.RequestSending.Providers.Attributive;
 using System.Net.Http.Json;
 using Xunit.Abstractions;
