+++ src/STrain.CQS.Test.Unit/Http/RequestSending/AttributivePathProviderTest.cs
