+++ src/STrain.CQS.Test.Unit/STrain.CQS.Test.Unit.csproj
@@ -9,6 +9,14 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Remove="Http\RequestSending\AttributiveBodyParameterProviderTest.cs" />
+    <Compile Remove="Http\RequestSending\AttributiveHeaderParameterProviderTest.cs" />
+    <Compile Remove="Http\RequestSending\AttributiveMethodProviderTest.cs" />
+    <Compile Remove="Http\RequestSending\AttributivePathProviderTest.cs" />
+    <Compile Remove="Http\RequestSending\AttributiveQueryParameterProviderTest.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="AutoFixture" Version="4.17.0" />
     <PackageReference Include="FluentValidation" Version="11.2.2" />
     <PackageReference Include="MartinCostello.Logging.XUnit" Version="0.3.0" />
