+++ src/STrain.Sample.Backend/Wireup/CQSWireUp.cs
@@ -25,11 +25,7 @@ namespace STrain.Sample.Backend.Wireup
                 if (request.GetType().Name.Contains("External")) return "External";
                 else return "Generic";
             },
-                builder =>
-                        {
-                            builder.AddHttpSender("External", (options, configuration) => configuration.Bind("Senders:External", options)).UseDefaults();
-                            builder.AddHttpSender("Generic", (options, configuraion) => configuraion.Bind("Senders:Generic", options)).UseGenericDefaults();
-                        });
+                builder => builder.AddHttpSender("Generic", (options, configuraion) => configuraion.Bind("Senders:Generic", options)));
         }
     }
 }
